Put the function inside utils folder for better organization
